restructuralisation and deletion of old files
